create UserBlock component
